Refactor JavaScript and CSS into modular structure
This refactoring improves code organization and maintainability without
introducing any frameworks or build tools. All code uses native browser
features (ES6 modules and CSS @import).

JavaScript Changes:
- Split monolithic app.js (1,177 lines) into 14 focused modules
- Created directory structure: core/, services/, features/, ui/
- Implemented EventBus for decoupled component communication
- Separated concerns: services handle API/WebSocket, features handle
  business logic, UI managers handle display/interaction
- Main modules:
  * Core: EventBus, SquareGolfApp
  * Services: ApiClient, WebSocketService, DeviceService, GSProService
  * Features: AlignmentManager, SettingsManager, CameraManager, ShotMonitor
  * UI: ToastManager, LoadingManager, ScreenManager
- Updated index.html to use <script type="module">

CSS Changes:
- Split monolithic style.css (1,587 lines) into 14 component files
- Organized by purpose: base, layout, components, features, responsive
- Main style.css now imports all modules
- Created files:
  * Base: variables.css, base.css
  * Layout: layout.css, status-bar.css
  * Components: cards.css, forms.css, buttons.css, status-elements.css,
    tabs.css, loading.css
  * Features: alignment.css, shot-monitor.css
  * Utilities: responsive.css

Benefits:
- Each file is now 15-330 lines (easy to navigate)
- Clear separation of concerns
- Better maintainability and testability
- Easier onboarding for new developers
- No build process required
- Uses only native browser features

Backup files (.backup) preserved for safety.
