+++ web/static/css/base.css
@@ -1,0 +1,14 @@
+/* Reset and base styles */
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+}
+
+body {
+    font-family: 'Inter', -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif;
+    font-size: var(--font-sm);
+    line-height: var(--line-height-relaxed);
+    color: var(--text-primary);
+    background: var(--bg-primary);
+}
