+++ web/static/css/variables.css
@@ -1,0 +1,171 @@
+/* CSS Variables */
+:root {
+    /* Colors */
+    --color-primary: #3182ce;
+    --color-primary-dark: #2563eb;
+    --color-primary-darker: #1d4ed8;
+    --color-success: #22c55e;
+    --color-success-dark: #16a34a;
+    --color-success-darker: #15803d;
+    --color-danger: #ef4444;
+    --color-danger-dark: #dc2626;
+    --color-danger-darker: #b91c1c;
+    --color-warning: #eab308;
+    --color-warning-dark: #d69e2e;
+    --color-warning-darker: #b7791f;
+
+    /* Grays */
+    --gray-50: #f7fafc;
+    --gray-100: #f8f9fa;
+    --gray-200: #edf2f7;
+    --gray-300: #e2e8f0;
+    --gray-400: #cbd5e0;
+    --gray-500: #a0aec0;
+    --gray-600: #718096;
+    --gray-700: #4a5568;
+    --gray-800: #2d3748;
+    --gray-900: #1a202c;
+
+    /* Semantic colors */
+    --bg-primary: var(--gray-50);
+    --bg-card: white;
+    --bg-sidebar: var(--gray-900);
+    --text-primary: var(--gray-800);
+    --text-secondary: var(--gray-700);
+    --text-muted: var(--gray-600);
+    --text-light: #cbd5e1;
+    --border-color: var(--gray-300);
+
+    /* Status colors */
+    --status-connected-bg: #c6f6d5;
+    --status-connected-text: #22543d;
+    --status-warning-bg: #fef5e7;
+    --status-warning-text: #b7791f;
+    --status-error-bg: #fed7d7;
+    --status-error-text: #742a2a;
+
+    /* Spacing */
+    --spacing-xs: 0.25rem;
+    --spacing-sm: 0.5rem;
+    --spacing-md: 0.75rem;
+    --spacing-lg: 1rem;
+    --spacing-xl: 1.5rem;
+    --spacing-2xl: 2rem;
+    --spacing-3xl: 3rem;
+
+    /* Border radius */
+    --radius-sm: 4px;
+    --radius-md: 6px;
+    --radius-lg: 8px;
+    --radius-xl: 12px;
+
+    /* Transitions */
+    --transition-fast: 0.2s ease;
+    --transition-normal: 0.3s ease;
+    --transition-slow: 0.4s ease;
+
+    /* Shadows */
+    --shadow-sm: 0 1px 3px rgba(0, 0, 0, 0.1);
+    --shadow-md: 0 4px 12px rgba(0, 0, 0, 0.1);
+    --shadow-lg: 0 4px 12px rgba(0, 0, 0, 0.15);
+    --shadow-xl: 0 6px 16px rgba(0, 0, 0, 0.2);
+
+    /* Font sizes */
+    --font-xs: 0.75rem;
+    --font-sm: 0.875rem;
+    --font-base: 1rem;
+    --font-lg: 1.125rem;
+    --font-xl: 1.25rem;
+    --font-2xl: 1.5rem;
+    --font-3xl: 2rem;
+
+    /* Font weights */
+    --weight-normal: 400;
+    --weight-medium: 500;
+    --weight-semibold: 600;
+    --weight-bold: 700;
+    --weight-extrabold: 800;
+    --weight-black: 900;
+
+    /* Icon sizes */
+    --icon-xs: 1rem;
+    --icon-sm: 1.125rem;
+    --icon-md: 1.25rem;
+    --icon-lg: 1.5rem;
+    --icon-xl: 1.75rem;
+    --icon-2xl: 2rem;
+
+    /* Border widths */
+    --border-width-thin: 1px;
+    --border-width-normal: 2px;
+    --border-width-thick: 4px;
+
+    /* Additional spacing values */
+    --spacing-2xs: 0.375rem;
+
+    /* Letter spacing */
+    --letter-spacing-tight: 0.02em;
+    --letter-spacing-normal: 0.05em;
+    --letter-spacing-wide: 0.1em;
+
+    /* Line heights */
+    --line-height-tight: 1;
+    --line-height-normal: 1.4;
+    --line-height-relaxed: 1.5;
+    --line-height-loose: 1.6;
+
+    /* Component dimensions */
+    --sidebar-width: 240px;
+    --compass-size: 240px;
+    --ball-viz-size: 280px;
+    --toast-max-width: 400px;
+    --spinner-size: 40px;
+    --handedness-btn-size: 48px;
+    --min-touch-target: 120px;
+    --info-grid-min: 200px;
+
+    /* Z-index layers */
+    --z-loading: 9999;
+    --z-toast: 10000;
+
+    /* Animation timing */
+    --timing-slow: 0.6s;
+    --timing-medium: 1s;
+    --timing-very-slow: 1.5s;
+
+    /* Responsive breakpoints */
+    --breakpoint-lg: 1200px;
+    --breakpoint-md: 768px;
+
+    /* Common shadow values */
+    --shadow-sidebar: 2px 0 4px rgba(0, 0, 0, 0.1);
+    --shadow-button-primary: 0 2px 4px rgba(49, 130, 206, 0.3);
+    --shadow-button-primary-hover: 0 4px 8px rgba(49, 130, 206, 0.4);
+    --shadow-button-success: 0 2px 4px rgba(34, 197, 94, 0.3);
+    --shadow-button-success-hover: 0 4px 8px rgba(34, 197, 94, 0.4);
+    --shadow-button-danger: 0 2px 4px rgba(239, 68, 68, 0.3);
+    --shadow-button-danger-hover: 0 4px 8px rgba(239, 68, 68, 0.4);
+    --shadow-hover-lift: 0 4px 8px rgba(0, 0, 0, 0.15);
+    --shadow-hover-primary: 0 4px 8px rgba(49, 130, 206, 0.2);
+
+    /* Focus ring */
+    --focus-ring-primary: 0 0 0 4px rgba(49, 130, 206, 0.15);
+    --focus-ring-danger: 0 0 0 4px rgba(239, 68, 68, 0.15);
+
+    /* Overlay opacity */
+    --overlay-dark: rgba(0, 0, 0, 0.5);
+    --overlay-light: rgba(0, 0, 0, 0.05);
+    --overlay-medium: rgba(0, 0, 0, 0.1);
+
+    /* Common gradient stops */
+    --gradient-angle: 135deg;
+
+    /* Other dimensions */
+    --max-content-width: 1200px;
+    --max-height-collapsible: 500px;
+    --ripple-size: 300px;
+    --border-thick-styled: 4px;
+    --border-radius-handedness: 10px;
+    --min-height-metrics: 200px;
+    --compass-mobile-size: 250px;
+}
