+++ web/static/js/services/WebSocketService.js
@@ -1,0 +1,69 @@
+// services/WebSocketService.js
+export class WebSocketService {
+    constructor(eventBus) {
+        this.ws = null;
+        this.reconnectInterval = null;
+        this.eventBus = eventBus;
+    }
+
+    connect() {
+        const protocol = window.location.protocol === 'https:' ? 'wss:' : 'ws:';
+        const wsUrl = `${protocol}//${window.location.host}/ws`;
+
+        try {
+            this.ws = new WebSocket(wsUrl);
+
+            this.ws.onopen = () => {
+                console.log('WebSocket connected');
+                this.eventBus.emit('ws:connected');
+                if (this.reconnectInterval) {
+                    clearInterval(this.reconnectInterval);
+                    this.reconnectInterval = null;
+                }
+            };
+
+            this.ws.onmessage = (event) => {
+                try {
+                    const message = JSON.parse(event.data);
+                    this.eventBus.emit('ws:message', message);
+                } catch (error) {
+                    console.error('Error parsing WebSocket message:', error);
+                }
+            };
+
+            this.ws.onclose = () => {
+                console.log('WebSocket disconnected');
+                this.eventBus.emit('ws:disconnected');
+                this.scheduleReconnect();
+            };
+
+            this.ws.onerror = (error) => {
+                console.error('WebSocket error:', error);
+                this.eventBus.emit('ws:error', error);
+            };
+        } catch (error) {
+            console.error('Failed to connect WebSocket:', error);
+            this.scheduleReconnect();
+        }
+    }
+
+    scheduleReconnect() {
+        if (this.reconnectInterval) return;
+
+        this.reconnectInterval = setInterval(() => {
+            console.log('Attempting to reconnect WebSocket...');
+            this.connect();
+        }, 3000);
+    }
+
+    disconnect() {
+        if (this.ws) {
+            this.ws.close();
+            this.ws = null;
+        }
+        if (this.reconnectInterval) {
+            clearInterval(this.reconnectInterval);
+            this.reconnectInterval = null;
+        }
+    }
+}
