+++ web/static/js/ui/ToastManager.js
@@ -1,0 +1,34 @@
+// ui/ToastManager.js
+export class ToastManager {
+    constructor() {
+        this.container = document.getElementById('toastContainer');
+    }
+
+    show(message, type = 'info') {
+        const toast = document.createElement('div');
+        toast.className = `toast ${type}`;
+        toast.textContent = message;
+
+        this.container.appendChild(toast);
+
+        setTimeout(() => {
+            toast.remove();
+        }, 5000);
+    }
+
+    success(message) {
+        this.show(message, 'success');
+    }
+
+    error(message) {
+        this.show(message, 'error');
+    }
+
+    warning(message) {
+        this.show(message, 'warning');
+    }
+
+    info(message) {
+        this.show(message, 'info');
+    }
+}
